refactor: LineGenerator static util 클래스로 변경
